ai_inference_api/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
fastapi==0.104.1
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h11==0.16.0
h5py==3.15.1
httptools==0.7.1
humanfriendly==10.0
idna==3.11
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
numpy==1.26.2
onnx==1.19.1
onnxruntime==1.23.1
opencv-python-headless==4.8.1.78
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
Pillow==10.1.0
protobuf==6.33.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
Pygments==2.19.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.31.0
rich==14.2.0
scikit-learn==1.3.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.20.0
termcolor==3.1.0
tf2onnx==1.16.1
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
